Change the user interface
Only use the /devices command, which will list devices as inline buttons.
When you click on the inline button, the device will be added/removed
from the black list.

Remove methods of the Session class that return device lists as a string,
as well as MAC address validation, since it is not used.Add filtering to
add and remove MAC addresses from the blacklist and logging.